Refactor Embeddings Management & Remove Business Logic From Controller
